fix: styling of images in links dialog
