+++ app/components/EmailPreviewHeader.tsx
@@ -7,21 +7,11 @@ import type { Email, User } from '@prisma/client';
 import {
   EnvelopeClosedIcon,
   ExternalLinkIcon,
-  Link2Icon,
   TrashIcon,
 } from '@radix-ui/react-icons';
 import { AlertDialog } from './AlertDialog';
 import { AlertDialogTrigger } from './ui/alert-dialog';
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from '~/components/ui/dialog';
-import { LINK_COPY_SUCCESS_MESSAGE } from '~/lib';
-import { useToast } from './ui/use-toast';
+import { HarvestLinks } from './HarvestLinks';
 
 interface EmailPreviewHeaderProps {
   email: Email;
@@ -34,19 +24,8 @@ export function EmailPreviewHeader({
   userId,
   setSelected,
 }: EmailPreviewHeaderProps) {
-  const { toast } = useToast();
   const submit = useSubmit();
   const [, setSearchParams] = useSearchParams();
-
-  const links = React.useMemo(() => {
-    if (typeof window === 'undefined') {
-      return [];
-    }
-    const parser = new DOMParser();
-    const doc = parser.parseFromString(email.html, 'text/html');
-    const links = Array.from(doc.querySelectorAll('a'));
-    return links;
-  }, [email.html]);
 
   return (
     <div className="max-h-48 overflow-hidden p-4 text-zinc-900">
@@ -121,54 +100,7 @@ export function EmailPreviewHeader({
             </Button>
           </Link>
         </Tooltip>
-
-        <Dialog>
-          <Tooltip content="Harvest links">
-            <DialogTrigger>
-              <Button variant="ghost" type="button" size="icon">
-                <Link2Icon />
-              </Button>
-            </DialogTrigger>
-          </Tooltip>
-          <DialogContent onCloseAutoFocus={(e) => e.preventDefault()}>
-            <DialogHeader>
-              <DialogTitle>Found the following links in the email</DialogTitle>
-              <DialogDescription>
-                Click a link to copy it to your clipboard
-              </DialogDescription>
-              {links.length > 0 ? (
-                <div className="grid auto-rows-fr grid-cols-2 grid-rows-6 place-items-center gap-4 pt-8">
-                  {links.map((link) => {
-                    return (
-                      <Tooltip
-                        key={link.href}
-                        content={link.href}
-                        sideOffset={4}
-                      >
-                        <div
-                          className="cursor-pointer text-center"
-                          dangerouslySetInnerHTML={{
-                            __html: link.innerHTML,
-                          }}
-                          onClick={async () => {
-                            await navigator.clipboard.writeText(link.href);
-                            toast({
-                              description: LINK_COPY_SUCCESS_MESSAGE,
-                            });
-                          }}
-                        />
-                      </Tooltip>
-                    );
-                  })}
-                </div>
-              ) : (
-                <DialogDescription>
-                  No links were found in the email
-                </DialogDescription>
-              )}
-            </DialogHeader>
-          </DialogContent>
-        </Dialog>
+        <HarvestLinks email={email} />
       </Form>
       <Separator />
     </div>
